// local-review-stack/anki-mcp-server/src/mcp/types/anki.types.ts
export interface AnkiCard {
  cardId: number;
  fields: Record<string, { value: string; order: number }>;
  fieldOrder: number;
  question: string;
  answer: string;
  modelName: string;
  ord: number;
  deckName: string;
  css: string;
  factor?: number;
  interval?: number;
  note: number;
  type: number;
  queue: number;
  due?: number;
  reps?: number;
  lapses?: number;
  left?: number;
  mod?: number;
  flags?: number;
  tags?: string[];
}